+++ data/tr2/ship/data/scripts/assault.lua
@@ -1,4 +1,7 @@
-trx.events.on_level_start(function(level)
+trx.events.on_game_start(function(level, is_save)
+  if is_save then
+    return
+  end
   local records = trx.assault_stats.list_records()
   if #records > 1 then
     trx.music.play(22)